[SOIN] Isole les types front liés aux tests dans des fichier .type.ts
