+++ front/lib-svelte/src/niveaux-maturite/IllustrationNiveauMaturite.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import type { NiveauMaturite } from './NiveauxMaturite.donnees';
+  import type { NiveauMaturite } from './NiveauxMaturite.type';
 
   export let niveauSelectionne: NiveauMaturite;
 </script>
