+++ front/lib-svelte/src/niveaux-maturite/NiveauxMaturite.donnees.ts
@@ -1,23 +1,4 @@
-export type IdNiveau =
-  | 'insuffisant'
-  | 'emergent'
-  | 'intermediaire'
-  | 'confirme'
-  | 'optimal';
-
-export type NiveauMaturite = {
-  id: IdNiveau;
-  label: string;
-  description: string;
-  priorite: string;
-  priseEnCompteRisque: string;
-  posture: string;
-  pilotage: string;
-  ressourcesHumaines: string;
-  budget: string;
-  adoptionSolutions: string;
-  niveauSecurite: string;
-};
+import type { IdNiveau, NiveauMaturite } from './NiveauxMaturite.type';
 
 export const niveauxMaturite: NiveauMaturite[] = [
   {
