+++ front/lib-svelte/src/test-maturite/TestMaturite.donnees.ts
@@ -1,32 +1,4 @@
-export type EtapeTest = {
-  titre: string;
-  question: string;
-  propositions: string[];
-  id: IdRubrique | 'infos-complementaires';
-};
-
-export type Question = {
-  titre: string;
-  question: string;
-  propositions: string[];
-  id: IdRubrique;
-};
-
-export type IdRubrique =
-  | 'pilotage'
-  | 'budget'
-  | 'ressources-humaines'
-  | 'adoption-solutions'
-  | 'prise-en-compte-risque'
-  | 'posture';
-
-export type Rubrique = {
-  id: IdRubrique;
-  label: string;
-  ancrageTexte: 'start' | 'end';
-  alignementVertical: 'middle' | 'auto' | 'hanging';
-  lettre: 'A' | 'B' | 'C' | 'D' | 'E' | 'F';
-};
+import type { EtapeTest, Question, Rubrique } from './TestMaturite.type';
 
 export const questions: Question[] = [
   {
@@ -107,8 +79,6 @@ export const questions: Question[] = [
     ],
   },
 ];
-
-export type ReponsesResultatTest = Record<IdRubrique, number>;
 
 export const etapesTestMaturite: EtapeTest[] = [
   ...questions,
